Make the progress bar play well with systemd.
Unfortunately, we won't see the updates directly out of
'sudo systemctl restart opennms'. To see status during OpenNMS
restarts use 'journalctl -f -u opennms' to monitor separately and/or
something like 'watch -n 1 systemctl status opennms' (I personally
like seeing the systemd view with the latter).

We do a few things in the systemd case:
1. If the 'opennms' script isn't running in terminal with a tty, the
   terminal.columns property isn't passed and Invoker.java treats the
   terminal as non-interactive. We use this to tweak our output in a
   few ways to work well with systemd:
   - Use '\n' instead of '\r' for progress updates, to avoid
     'systemd status' from just reporting very unhelpful
     '[2.9K blob data]' messages.
   - Use ProgressBarStyle.ASCII only in non-interactive mode, and use
     the default style in interactive mode.
   - Disable continuous update in non-interactive mode because it makes
     the status output needlessly verbose. systemd alreday monitors
     whether the process is still alive, so the only value continous
     updates provide is to see the estimated remaining time update,
     which feels a lot less useful given the amount of 'systemd status'
     noise it generates.
2. Some moderate tweaks to 'opennms' when the progress bar is enabled
   and we are running in systemd:
   - Redirect opennms output to output.log just like we do normally
     when running outside of systemd.
   - Run tail(1) with "--pid $pid" to that tail(1) will die when the
     OpenNMS process exits. Without this systemd will still think the
     service is active. Also add a check in systemd mode to make sure
     that tail(1) supports "--pid", and if not, disable the progress
     bar. See the notes at the end of the commit for what happens if
     we don't do this.
   - When stopping, don't run another tail(1) process because the one
     from the start will still be running in the CGroup.
3. Pass "-s" to the stop command in the systemd opennms.service unit
   file to support the last item above to avoid an extra tail(1).
4. Tweak our invocations of tail(1) to use "-F -n 0" so we work with
   more tail(1) variants (in particular, the one in Ubuntu).
5. Increase our progressbar.log rotation size from 10 kB to 100 kB
   because tail(1) on Ubuntu spits out a log message or two when the
   file is rotated and these messages can't be disabled. :(

Also, quiet down tput(1) when there is no tty present.

=== Okay, so about this systemd forking business and tail --pid ===

Currently in systemd mode and with these changes, we have an
interesting and slightly problematic process structure:

     CGroup: /system.slice/opennms.service
             ├─29777 bash /usr/share/opennms/bin/opennms -s start
             ├─29778 tail -F -n 0 --pid 29777 /var/log/opennms/progressbar.log
             └─29779 /usr/lib/jvm/java-11-openjdk-amd64/bin/java ... -jar /usr/share/opennms/lib/opennms_bootstrap.jar start

(Note: the only difference without this change would be the lack of the
tail process).

Note that we have the bash process for 'opennms -s start' that sticks
around after OpenNMS has been started. The 'opennms' script runs
'runCmd "${CMD[@]}" &' and 'runCmd' is a shell function not an external
command (although inside 'runCmd' it does call an external command).
Since that shell function is what is backgrounded, a copy of the shell
sticks around to run the command (in the foreground in that shell
function to make things even more confusing, so I think this happens in
all cases of running OpenNMS using this script).

You'll also notice that it is the PID for that bash process which is
passed to 'tail --pid' (29777 in the above sample) *not* the OpenNMS
process itself (29779 in this case). Luckily, when OpenNMS exits, the
shell should exit, as well.

If we don't use "tail --pid" and OpenNMS exits abruptly (without
'opennms stop' being called **by systemd**), we end up with just the
tail(1) process running in the CGroup:

     CGroup: /system.slice/opennms.service
             └─163703 tail -F -n 0 /var/log/opennms/progressbar.log

And systemd thinks things are still happy, which is **bad**:

     Active: active (running) since Sat 2022-10-22 02:37:54 UTC; 5min ago

So, that's why the startup script is picky about 'tail --pid' being
there in systemd mode with the progress bar, and if it isn't, the
progress bar feature is disabled.

We should probably eventually switch to use Type=notify in systemd
instead of Type=forking that we have now. Ideally, we'd just exec
the OpenNMS Java process on startup and have it notify systemd when
it is up and running. See:
https://www.freedesktop.org/software/systemd/man/systemd.service.html#Type=

Here are some references for doing the notification directly in Java:
- https://github.com/faljse/SDNotify
- https://gist.github.com/juur/048cc3d0554953b717e9c6867970f30e
- Once we require Java 16, it has Unix domain socket channels:
  https://inside.java/2021/02/03/jep380-unix-domain-sockets-channels/
- Netty has it now: https://netty.io/4.0/api/io/netty/channel/unix/DomainSocketChannel.html

As an FYI, here are two other things I noticed:

Oct 23 00:46:18 dgregor-vm-perf systemd[1]: opennms.service: Can't open PID file /run/opennms/opennms.pid (yet?) after start: Operation not permitted
Oct 23 00:46:19 dgregor-vm-perf systemd[1]: opennms.service: Supervising process 37802 which is not our child. We'll most likely not notice when it exits.

I think the latter message about supervising is related to using the
forking configuration. Both of these seem to happen with or without
the changes in this commit.